Adds code to switch to appropriate wifi channel.

+++ soloutils/wifi.py
@@ -5,6 +5,102 @@ from distutils.version import LooseVersion
 # and then executing it, so we are resilient to network dropouts.
 
 SCRIPT = """
+cat > /tmp/align_channel.py <<'SCRIPT'
+import itertools, re, subprocess
+from pprint import pprint
+import sys
+
+# align_channel.py 3DR
+
+if len(sys.argv) < 2:
+    print 'Usage: align_channel.py <ssid>'
+    sys.exit(1)
+
+def build_tree(data):
+    lines = data.split('\\n')
+
+    stack = [[]]
+    indent = ['']
+    i = 0
+    while i < len(lines):
+        line = lines[i]
+        if line[:len(indent[0])] == indent[0]:
+            leadtab = line[len(indent[0]):len(indent[0]) + 1]
+            if leadtab == '\\t':
+                stack.insert(0, [])
+                indent.insert(0, indent[0] + leadtab)
+            sub = line[len(indent[0]):]
+            if len(sub):
+                stack[0].append(line[len(indent[0]):])
+            i += 1
+        else:
+            last = stack.pop(0)
+
+            # Tabs can roll onto previous line if short enough
+            if stack[0][-1] and '\\t' in stack[0][-1]:
+                a = stack[0][-1].split('\\t')
+                stack[0][-1] = a[0] or ''
+                last.insert(0, a[1] or '')
+
+            stack[0].append(last)
+            indent.pop(0)
+
+    return stack.pop()
+
+input = subprocess.check_output('iw dev wlan0 scan ap-force', shell=True)
+networks = []
+for entry in build_tree(input):
+    if not isinstance(entry, basestring):
+        ssid = None
+        freq = None
+        for e in entry:
+            if isinstance(e, basestring):
+                if 'SSID' in e:
+                    try:
+                        ssid = e.split(None, 2)[1]
+                    except:
+                        pass
+                if 'freq' in e:
+                    try:
+                        freq = e.split(None, 2)[1]
+                    except:
+                        pass
+        networks.append((ssid, freq))
+
+print networks
+
+input = subprocess.check_output('iw dev', shell=True)
+curfreq = None
+noht = None
+for phy in build_tree(input):
+    if not isinstance(phy, basestring):
+        for i in xrange(0, len(phy)):
+            if isinstance(phy[i], basestring) and 'wlan0-ap' in phy[i]:
+                try:
+                    curfreq = re.search(r'(\d+) [Mm][Hh][Zz]', '\\n'.join(phy[i+1])).group(1)
+                except:
+                    pass
+                noht = re.search(r'[Nn][Oo] HT', '\\n'.join(phy[i+1]))
+
+if not curfreq:
+    print 'Could not identify current frequency, trying anyway...'
+
+targetfreq = None
+for (S, M) in networks:
+    if S == sys.argv[1]:
+        if not targetfreq or targetfreq != curfreq:
+            targetfreq = M
+
+if not targetfreq:
+    print 'Specified SSID not found. Are you sure it was typed correctly?'
+    sys.exit(1)
+
+if curfreq != targetfreq:
+    input = subprocess.check_output('hostapd_cli chan_switch 1 ' + str(targetfreq) + (' ht' if not noht else ''), shell=True)
+
+print '(Target frequency matched.)'
+SCRIPT
+
 cat > /tmp/setupwifi.sh << 'SCRIPT'
 
 wget -O- http://example.com/ --timeout=5 >/dev/null 2>&1
@@ -24,6 +120,9 @@ EOF
     /etc/init.d/dnsmasq restart
     sleep 2
 
+    ifconfig wlan0 up
+    python /tmp/align_channel.py {ssid}
+
     wpa_supplicant -i wlan0 -c /etc/wpa_client.conf -B
     udhcpc -i wlan0
 
@@ -36,12 +135,6 @@ EOF
 fi
 
 echo 'connecting to the internet...'
-
-smart query --installed | grep 'kernel-module-iptable-filter' >/dev/null 2>&1
-if [ $? != 0 ]; then
-    python -c "import urllib2, sys; sys.stdout.write(urllib2.urlopen('https://s3.amazonaws.com/solo-packages/controller/kernel-module-iptable-filter-3.10.17-r0.imx6solo_3dr_artoo.rpm').read()); sys.stdout.flush()" > /tmp/iptable_filter.rpm
-    rpm -iv --replacepkgs /tmp/iptable_filter.rpm
-fi
 insmod /lib/modules/3.10.17-rt12-*/kernel/net/ipv4/netfilter/iptable_filter.ko
 
 iptables -t nat -A POSTROUTING -o wlan0 -j MASQUERADE
@@ -58,13 +151,13 @@ def main(args):
     controller = soloutils.connect_controller(await=True)
 
     controller_version = soloutils.controller_versions(controller)['version']
-    if LooseVersion('1.1.15') != LooseVersion(controller_version):
-        print 'error: expecting Controller to be EXACTLY EQUAL TO 1.1.15'
+    if LooseVersion('1.2.0') > LooseVersion(controller_version):
+        print 'error: expecting Controller to be >= 1.2.0'
         print 'your Controller version: {}'.format(controller_version)
-        print 'please flash your Controller with 1.1.15 to run this command.'
+        print 'please flash your Controller with a newer version to run this command.'
         print ''
-        print '    solo update solo 1.1.15'
-        print '    solo update controller 1.1.15'
+        print '    solo update solo latest'
+        print '    solo update controller latest'
         sys.exit(1)
 
     code = soloutils.command_stream(controller, SCRIPT.format(ssid=args['--name'], password=args['--password']))
